BUGFIX [#159579536] Apple iOS support for web xj player  \   * [ui/player-ui]       + Enhanced web audio unlocking on touch event       + JSFiddle https://jsfiddle.net/charneykaye/jm847Lr0/ works OK from iOS device!       + Deprecate proprietary BinaryResource in favor of simple method       + Don't bother with echoing JSON response from API, we can confirm that ourselves.       + Do echo all messages from player and segment-audio       + It was necessary to implement CORS perfectly, because iOS is strict       + Use legacy WebAudio API to decode audio
[#166556739] JSR-310 specified ISO-8601 compatible time
 \
  * [core/time]
      + Deprecate util/TimestampUTC and use java.time everywhere
      + Deprecate `sql.Timestamp` everywhere in favor of `Instant` -- except where actual SQL interface occurs
      + API uses Instant.truncateTo(ChronoUnit.MICROS) for microsecond precision
      + When transmogrifying Model<>Record SQL Timestamp is used
      + Modify all tests to send and expect time as ISO-8601 text

[#160112750] XJ player uses only core ECMA6
 \
  * [ui/player-ui]
      + DOM utility provides the only few things we ever needed from jquery
      + Deprecate jQuery

Adhoc:
  * [ui/main-site]
      + Upgrade gulpfile for Gulp 4
      + Update (tm) to new (R) registered trademarks

+++ hub/src/main/java/io/xj/hub/resource/segment/SegmentIndexResource.java
@@ -49,7 +49,7 @@ public class SegmentIndexResource extends HubResource {
   BigInteger fromOffset;
 
   @QueryParam("fromSecondsUTC")
-  BigInteger fromSecondsUTC;
+  Long fromSecondsUTC;
 
   /**
    Get all segments.
